cambios para los reportes

+++ LittleWishWebbackup/src/main/java/pe/edu/upc/aaw/littlewishproject/repositories/ICertificacionesRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.aaw.littlewishproject.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.aaw.littlewishproject.entities.Certificaciones;
 
-public interface ICertificacionesRepository extends JpaRepository <Certificaciones, Long> {
+public interface ICertificacionesRepository extends JpaRepository <Certificaciones, Integer> {
 }
