+++ LittleWishWebbackup/src/main/java/pe/edu/upc/aaw/littlewishproject/repositories/IIdiomasRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.aaw.littlewishproject.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.aaw.littlewishproject.entities.Idiomas;
 
-public interface IIdiomasRepository extends JpaRepository<Idiomas, Long> {
+public interface IIdiomasRepository extends JpaRepository<Idiomas, Integer> {
 }
